+++ src/lib.rs
@@ -1,4 +1,6 @@
 pub mod commands;
 pub mod templates;
 pub mod auth;
-pub mod utils;
+
+mod utils;
+pub use utils::*;